Version 0.3.0

+++ cmd/install/install.go
@@ -2,6 +2,7 @@ package install
 
 import (
 	_ "embed"
+	"errors"
 	"ismartcoding/plainnas/internal/consts"
 	"log"
 	"os"
@@ -18,11 +19,20 @@ var configToml []byte
 
 // Removed worker service and build script embedding (no C worker).
 
-func Install() {
+type Options struct {
+	WithLibreOffice bool
+}
+
+func Install(opts Options) {
 	// Installer output should be easy to read.
 	log.SetFlags(0)
 
 	installAllDeps()
+	if opts.WithLibreOffice {
+		installLibreOffice()
+	} else {
+		printNote("DOC/DOCX preview (LibreOffice)", "Optional. Run: plainnas install --with-libreoffice")
+	}
 	runSelfChecks()
 
 	// Ensure directories exist (systemd: 0755, app config: 0755)
@@ -33,12 +43,21 @@ func Install() {
 		log.Println(err)
 	}
 
-	// Replace id in configToml with a new uuid before writing
-	newID := uuid.NewString()
-	re := regexp.MustCompile(`id = ".*"`)
-	configStr := string(configToml)
-	configStr = re.ReplaceAllString(configStr, "id = \""+newID+"\"")
-	if err := os.WriteFile(consts.ETC_MAIN_CONFIG, []byte(configStr), 0640); err != nil {
+	// Do not overwrite user config. Only write default config on first install.
+	if _, err := os.Stat(consts.ETC_MAIN_CONFIG); err == nil {
+		printOK("Config", "Keeping existing "+consts.ETC_MAIN_CONFIG)
+	} else if errors.Is(err, os.ErrNotExist) {
+		// Replace id in configToml with a new uuid before writing
+		newID := uuid.NewString()
+		re := regexp.MustCompile(`id = ".*"`)
+		configStr := string(configToml)
+		configStr = re.ReplaceAllString(configStr, "id = \""+newID+"\"")
+		if err := os.WriteFile(consts.ETC_MAIN_CONFIG, []byte(configStr), 0640); err != nil {
+			log.Println(err)
+		} else {
+			printOK("Config", "Written "+consts.ETC_MAIN_CONFIG)
+		}
+	} else {
 		log.Println(err)
 	}
 }
